services/auth-service/requirements.txt
fastapi==0.111.0
bcrypt==4.0.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
httpx==0.27.0
